requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0

# Fine-tuning
bitsandbytes>=0.43.0
accelerate>=0.27.0
peft>=0.10.0

# API serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Demo
gradio>=4.0.0

# Tracking
mlflow>=2.8.0
wandb>=0.16.0

# Utilities
Pillow>=10.0.0
pyyaml>=6.0
requests>=2.31.0
numpy>=1.24.0

# Testing & code quality
pytest>=7.4.0
ruff>=0.1.0
black==23.12.1

# HuggingFace deployment
huggingface_hub>=0.19.0